@@ -1,10 +1,21 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client'; // ✅ new import for React 18+
 import App from './components/App';
 import { Provider } from 'react-redux';
 import thunk from 'redux-thunk';
-import { createStore,applyMiddleware } from 'redux';
+import { createStore, applyMiddleware } from 'redux';
+import { rootReducer } from './reducers';
 
+const store = createStore(rootReducer, applyMiddleware(thunk));
 
+console.log(store.getState());
 
-ReactDOM.render(<App/>, document.querySelector('#root'));
+const rootElement = document.querySelector('#root');
+if (rootElement) {
+  const root = ReactDOM.createRoot(rootElement); // ✅ React 18+
+  root.render(
+    <Provider store={store}>
+      <App />
+    </Provider>
+  );
+}
